% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_dict_entries}
\alias{add_dict_entries}
\title{add_dict_entries}
\usage{
add_dict_entries(dict, data, name_col, ...)
}
\arguments{
\item{dict}{data.frame to be added-to}

\item{data}{data.frame to be described}

\item{name_col}{string, column name}

\item{...}{additional entries, i.e. key = "value"}
}
\value{
data.frame
}
\description{
generate dictionary entries
}
\details{
returns a data-frame with n + 2 observations of three variables: name, attribute, value,
where n is the number of additional arguments
}

