% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coded_options.R
\name{coded_options}
\alias{coded_options}
\title{Convert an OCA style data dictionary into a long-form data.frame containing
values and labels of "Coded list" type variables}
\usage{
coded_options(dict)
}
\arguments{
\item{dict}{A data frame reflecting a data dictionary to validate}
}
\value{
A data.frame
}
\description{
Convert an OCA style data dictionary into a long-form data.frame containing
values and labels of "Coded list" type variables
}
