% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_from_data.R
\name{dict_from_data}
\alias{dict_from_data}
\title{Generate a data dictionary template from a dataset}
\usage{
dict_from_data(x, factor_threshold = 10, factor_values = c("int", "string"))
}
\arguments{
\item{x}{A data frame reflecting the dataset from which to generate a data
dictionary template}

\item{factor_threshold}{An integer representing the maximum number of unique
values within a dataset column for that column to be classified as a
factor-type (i.e. coded list) variable. Columns that are not recognized as
other data types (such as Numeric, Date, etc.) and have more than
\code{factor_threshold} unique values will be specified as type "Free text".
Defaults to 10.}

\item{factor_values}{Should values of factor-type (i.e. Coded list) variables
be generated as integers ("int") or strings ("string"). Defaults to "int".\preformatted{ label    | int | string
-------------------------
 Yes      | 0   | yes
 No       | 1   | no
 Not sure | 2   | not_sure
}}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame representing a data dictionary
template formatted to the OCA data sharing standard
}
\description{
Generate a data dictionary template from a dataset
}
\examples{
# read example dataset
path_data <- system.file("extdata", package = "datadict")
dat <- readxl::read_xlsx(file.path(path_data, "linelist_cleaned.xlsx"))

# generate data dictionary template from dataset
dict_from_data(dat, factor_values = "string")

}
