% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_from_redcap.R
\name{dict_from_redcap}
\alias{dict_from_redcap}
\title{Convert a REDCap data dictionary to the OCA data sharing standard}
\usage{
dict_from_redcap(x)
}
\arguments{
\item{x}{A data.frame reflecting a REDCap data dictionary}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame representing a data dictionary
formatted to the OCA data sharing standard
}
\description{
Convert a REDCap data dictionary to the OCA data sharing standard
}
\examples{
# path to example REDCap data dictionary
path_data <- system.file("extdata", package = "datadict")
path_redcap_dict <- file.path(path_data, "REDCapDataDictionaryDemo.csv")

# read dictionary
redcap_dict <- read.csv(path_redcap_dict)

# convert to OCA format
dict_from_redcap(redcap_dict)

}
