% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_coded_options.R
\name{generate_coded_options}
\alias{generate_coded_options}
\title{Generate a compact specification of the options for a variable of type 'Coded
list'}
\usage{
generate_coded_options(x, type_of_values = c("int", "string", "names"))
}
\arguments{
\item{x}{A vector of labels}

\item{type_of_values}{Type of values to generate corresponding to each
possible label. Either integers ("int"), strings ("string"), or take values
from the element names of \code{x} ("names").}
}
\value{
A string giving the possible values and corresponding labels for a given
variable, in compact format
}
\description{
The compact form is a single string with the value and label corresponding to
a given option separated with a comma, and options separated with a vertical
bar, e.g. "value1, Label 1 | value2, Label 2 | value3, Label 3"
}
\examples{
x <- c("Value #1", "Value #2", "Value #3")
names(x) = c("val1", "val2", "val3")

# use specified names as values
generate_coded_options(x, type_of_values = "names")

# generate integer values
generate_coded_options(x, type_of_values = "int")

# generate string values
generate_coded_options(x, type_of_values = "string")

}
