% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dict.R
\name{generate_dict}
\alias{generate_dict}
\title{Generate a data dictionary}
\usage{
generate_dict(
  data,
  columns = NULL,
  n_max = 10L,
  label_fallback = "NO QUESTION LABEL",
  label_width = 100L
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{columns}{An optional character vector of column names to include in
the dictionary. If \code{NULL} (default), information for all non-empty columns/
variables are included in the dictionary.}

\item{n_max}{An integer that controls how detailed numeric reporting should
be. If a numeric variable has more unique values than this threshold, it is
summarized  as a range (e.g., \verb{1 – 100}). If it has fewer, each unique value
receives its own row in the dictionary. Defaults to \code{10}.}

\item{label_fallback}{The placeholder text to use when a variable does not
have a defined label attribute. Defaults to \code{"NO QUESTION LABEL"}.}

\item{label_width}{An integer to truncate long variable labels for better
readability. Defaults to \code{100} characters.}
}
\value{
A tibble with the following structural metadata:
\itemize{
\item variable_name: The column/variable name as it appears in the data source.
\item variable_label: The descriptive variable label or the default placeholder.
\item variable_type: The type of data contained within the column/variable (e.g.,
numeric, character, labelled, factor). Data type is derived using R's \code{class()}
function.
\item variable_values: Individual data values or range of data values.
\item value_labels: The descriptive labels assigned to specific data values.
\item n_size: The frequency count for each value/label pair.
\item is_range: A boolean flag indicating if the \code{variable_values} column
represents a summarized range rather than a discrete value.
}
}
\description{
\code{generate_dict()} builds a data dictionary from a data frame
or tibble. It produces a structured overview of variable metadata, including
variable names, labels, data types, value distributions, and any available
values or value labels.
}
\examples{
generate_dict(data = nlsy)

# example shortening variable labels to 25 characters
generate_dict(data = gss, label_width = 25)

# example selecting a subset of variables from a tibble
generate_dict(
  data = gss,
  columns = c("year", "age", "sexnow1", "hispanic", "race", "marital")
)

}
\author{
Ama Nyame-Mensah
}
