% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_dict_wide}
\alias{get_dict_wide}
\title{get_dict_wide}
\usage{
get_dict_wide(dict, by_num = TRUE, print_hash = TRUE)
}
\arguments{
\item{dict}{}

\item{by_num}{logical (default TRUE)}

\item{print_hash}{logical (default TRUE)}
}
\value{
data.frame
}
\description{
print out a dictionary
}

