% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_entries}
\alias{get_entries}
\title{get_entries}
\usage{
get_entries(dict, by_num = TRUE, print_hash = TRUE)
}
\arguments{
\item{dict}{data.frame, dictionary}

\item{by_num}{logical (default TRUE)}

\item{print_hash}{logical (default TRUE)}
}
\value{
data.frame
}
\description{
print out a dictionary
}

