% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_anonymity_counts.R
\name{k_anonymity_counts}
\alias{k_anonymity_counts}
\title{Count the number of observations across unique combinations of indirect
identifiers within a dataset}
\usage{
k_anonymity_counts(x, vars, threshold = NULL)
}
\arguments{
\item{x}{A data frame}

\item{vars}{A character vector containing the name(s) of the variable(s) in
\code{x} to be included in the k-anonymity calculation}

\item{threshold}{Integer threshold indicating the minimum acceptable value of
k. Combinations with values of k below the threshold will be flagged and
returned. A return with 0 rows indicates that no combinations have values
of k below the threshold.}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame containing counts of unique
combinations of the variables specified in argument \code{vars}. If argument
\code{threshold} is specified, only the combinations with counts lower than the
threshold are returned, if any (i.e. combinations that do not meet the
specified value of k-anonymity).
}
\description{
Given a dataset and set of one or more variables that may be indirect
identifiers, the function returns a table of counts of the number of
observations corresponding to each unique combination of those variables
(i.e. k), optionally filtered to those combinations that do not meet the
user-specific threshhold of k-anonymity.
}
\examples{
# read example dataset
path_data <- system.file("extdata", package = "datadict")
dat <- readxl::read_xlsx(file.path(path_data, "linelist_cleaned.xlsx"))

# display combinations of gender and age_cat with k < 5
k_anonymity_counts(dat, vars = c("gender", "age_cat"), threshold = 5)

}
