% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_dict_entries}
\alias{make_dict_entries}
\title{make_dict_entries}
\usage{
make_dict_entries(data, name_col, ...)
}
\arguments{
\item{data}{data.frame to be described}

\item{name_col}{string, column name}

\item{...}{additional entries, i.e. key = "value"}
}
\value{
data.frame
}
\description{
generate dictionary entries
}
\details{
returns a data-frame with n + 2 observations of three variables: name_col, attribute, value,
where n is the number of additional arguments
}

