% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_entries_dict}
\alias{make_entries_dict}
\title{make_entries_dict}
\usage{
make_entries_dict(dict, data)
}
\arguments{
\item{dict}{data.frame with dictionary entries, must have column name_col}

\item{data}{data.frame to be described}
}
\value{
data.frame
}
\description{
sample_text
}

