% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass_data.R
\name{reclass_data}
\alias{reclass_data}
\title{Reclass the columns of a dataset to match the variable types specified in the
corresponding data dictionary}
\usage{
reclass_data(
  data,
  dict,
  format_date = "\%Y-\%m-\%d",
  format_time = "\%H:\%M:\%S",
  format_datetime = NULL
)
}
\arguments{
\item{data}{A data frame reflecting a dataset to be shared}

\item{dict}{A data frame reflecting the corresponding data dictionary}

\item{format_date}{Expected format for date variables. Defaults to
"\%Y-\%m-\%d".}

\item{format_time}{Expected format for date variables. Defaults to
"\%H:\%M:\%S".}

\item{format_datetime}{Expected format for date variables. Defaults to \code{NULL}
to use defaults in \link[lubridate:as_date]{lubridate::as_datetime}.}
}
\value{
The dataset with columns reclassed to match the data dictionary
}
\description{
Reclass the columns of a dataset to match the variable types specified in the
corresponding data dictionary
}
