% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_data.R
\name{valid_data}
\alias{valid_data}
\title{Check that a dataset is consistent with its corresponding data dictionary}
\usage{
valid_data(
  data,
  dict,
  format_date = "\%Y-\%m-\%d",
  format_time = "\%H:\%M:\%S",
  format_datetime = NULL,
  format_coded = "label",
  vals_withheld = NA,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame reflecting a dataset to be shared}

\item{dict}{A data frame reflecting the corresponding data dictionary}

\item{format_date}{Expected format for date variables. Defaults to
"\%Y-\%m-\%d".}

\item{format_time}{Expected format for date variables. Defaults to
"\%H:\%M:\%S".}

\item{format_datetime}{Expected format for date variables. Defaults to \code{NULL}
to use defaults in \link[lubridate:as_date]{lubridate::as_datetime}.}

\item{format_coded}{Expected format for coded-list variables, either "value"
or "label". Defaults to "label.}

\item{vals_withheld}{Expected value(s) in columns that are withheld. Default
to \code{NA}.}

\item{verbose}{Logical indicating whether to give warning describing the
checks that have failed. Defaults to TRUE.}
}
\value{
\code{TRUE} if all checks pass, \code{FALSE} if any checks fail
}
\description{
Includes the following checks:
\itemize{
\item variables marked 'withheld' in dictionary contain only missing values
\item all variables in dataset defined in dictionary
\item all variables defined in dictionary present in dataset
\item variables of type 'Logical' are all TRUE/FALSE (or 1/0)
\item variables of type 'Numeric' are valid numbers
\item variables of type 'Date' are valid dates
\item variables of type 'Time' are valid times
\item variables of type 'Datetimes' are valid date-times
\item variables of type 'Coded list' contain only allowed options
}
}
\examples{
# read example dataset
path_data <- system.file("extdata", package = "datadict")
dat <- readxl::read_xlsx(file.path(path_data, "linelist_cleaned.xlsx"))

# generate data dictionary template from dataset
dict <- dict_from_data(dat, factor_values = "string")

# dictionary column 'indirect_identifier' must be manually specified (yes/no)
dict$indirect_identifier <- "no"

# check for validity
valid_data(dat, dict)

}
