% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadict-package.R
\docType{package}
\name{datadict-package}
\alias{datadict}
\alias{datadict-package}
\title{datadict: Package For Generating a Data Dictionary}
\description{
A simple package for generating a data dictionary.
}
\author{
\strong{Maintainer}: Ama Nyame-Mensah \email{ama@anyamemensah.com}

}
\keyword{internal}
