% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadict-package.R
\docType{package}
\name{datadict-package}
\alias{datadict}
\alias{datadict-package}
\title{datadict: Generate a Data Dictionary}
\description{
A focused toolset for creating and managing data dictionaries. Allows users to document datasets by extracting variable labels and distributions, restructure complex metadata by expanding delimited entries into tidy formats, and verify the integrity of dictionary mappings.
}
\author{
\strong{Maintainer}: Ama Nyame-Mensah \email{ama@anyamemensah.com}

}
\keyword{internal}
