% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_anonymity.R
\name{k_anonymity}
\alias{k_anonymity}
\title{Check the k-anonymity of one or more variables in a dataset}
\usage{
k_anonymity(x, vars, threshold)
}
\arguments{
\item{x}{A data frame}

\item{vars}{A character vector containing the name(s) of the variable(s) in
\code{x} to be included in the k-anonymity calculation}

\item{threshold}{Integer threshold for k at or below which combinations
should be flagged (i.e. returned).}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}}-style data frame containing counts of the
variables specified in argument \code{vars} with counts (i.e. values of k) less
than or equal to \code{threshold}.
}
\description{
Given a dataset and set of one or more variables of interest (e.g. variables
that are potential indirect identifiers), the function returns the
combination(s) of those variables yielding counts (i.e. "k") less than or
equal to a user-specified threshold.
}
\examples{
# read example dataset
path_data <- system.file("extdata", package = "datadict")
dat <- readxl::read_xlsx(file.path(path_data, "linelist_cleaned.xlsx"))

# check whether there are combinations of gender and age_cat with k <= 5
k_anonymity(dat, vars = c("gender", "age_cat"), threshold = 5)

}
