% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_data.R
\name{valid_data}
\alias{valid_data}
\title{Check that a dataset is consistent with its corresponding data dictionary}
\usage{
valid_data(
  data,
  dict,
  format_date = "\%Y-\%m-\%d",
  format_time = "\%H:\%M:\%S",
  format_datetime = NULL,
  format_coded = "label"
)
}
\arguments{
\item{data}{A data frame reflecting a dataset to be shared}

\item{dict}{A data frame reflecting the corresponding data dictionary}

\item{format_date}{Expected format for date variables. Defaults to
"\%Y-\%m-\%d".}

\item{format_time}{Expected format for date variables. Defaults to
"\%H:\%M:\%S".}

\item{format_datetime}{Expected format for date variables. Defaults to \code{NULL}
to use defaults in \link[lubridate:as_date]{lubridate::as_datetime}.}

\item{format_coded}{Expected format for coded-list variables, either "value"
or "label". Defaults to "label.}
}
\value{
NULL. Prints checks and their results to the console.
}
\description{
Includes the following checks:
\itemize{
\item all variables in dataset defined in dictionary
\item all variables defined in dictionary present in dataset
\item variables of type 'Numeric' are valid numbers
\item variables of type 'Date' are valid dates
\item variables of type 'Time' are valid times
\item variables of type 'Datetimes' are valid date-times
\item variables of type 'Coded list' contain only allowed options
}
}
\examples{
# read example dataset
path_data <- system.file("extdata", package = "datadict")
dat <- readxl::read_xlsx(file.path(path_data, "linelist_cleaned.xlsx"))

# generate data dictionary template from dataset
dict <- dict_from_data(dat, factor_values = "string")

# check for validity
valid_data(dat, dict)

}
