% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_dict.R
\name{valid_dict}
\alias{valid_dict}
\title{Check that a data dictionary complies with the OCA data sharing standard}
\usage{
valid_dict(dict)
}
\arguments{
\item{dict}{A data frame reflecting a data dictionary to validate}
}
\value{
\code{TRUE} if all checks pass, else throws an error
}
\description{
Includes the following checks:
\itemize{
\item contains required columns
\item no missing variable names
\item no duplicated variable names
\item no missing data types
\item no non-valid data types
\item no missing choices (for coded-list type variables)
\item no non-valid choices (for coded-list type variables)
\item no non-valid values in column \code{origin}
\item no non-valid values in column \code{status}
}
}
\examples{
# read example dataset
path_data <- system.file("extdata", package = "datadict")
dat <- readxl::read_xlsx(file.path(path_data, "linelist_cleaned.xlsx"))

# generate data dictionary template from dataset
dict <- dict_from_data(dat, factor_values = "string")

# check for validity
valid_dict(dict)

}
